+++ packages/spor-react/src/loader/SkeletonText.tsx
@@ -1,7 +1,8 @@
-import { BoxProps, SkeletonText as ChakraSkeletonText } from "@chakra-ui/react";
+import { SkeletonText as ChakraSkeletonText, SkeletonTextProps as ChakraSkeletonTextProps } from "@chakra-ui/react";
 import React from "react";
 
-export type SkeletonTextProps = BoxProps;
+export type SkeletonTextProps = ChakraSkeletonTextProps
+  
 /**
  * SkeletonText renders a loading animation for a given text. It works great as a placeholder to avoid layout shifts.
  */